topic_intelligence_engine/requirements.txt
alembic>=1.12.0
psycopg2-binary>=2.9.9
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.1
sentence-transformers>=2.2.2
SQLAlchemy>=2.0.21
tenacity>=8.2.3
apify-client>=1.6.0
APScheduler>=3.10.0
openpyxl>=3.1.2
